* feat: make default label translatable

* i18n: add translation entries

* fix: remap unexist material

* i18n: add default stock name entries

* feat: make label translatable

* feat: optimize multiblock creation

* [Feature] Optimize CubicCurve (#1548)

* fix: remove potential overlapping points

* feat: add length for CubicCurve

* feat: optimize CubicCurve

* refactor: code cleanup

* refactor: code cleanup

* fix: incorrect calculate

* fix: insufficient track count

* fix: invalid count when Curve p1 equals to p2

* feat: improve precision algorithm

* fix: scale rail piece up a little bit

* fix: track not 1m long won't work

* feat: add range property to animatrix definition for better mapping (#1550)

* docs: update AUTHORS (#1560)

* fix: stock not removed (#1561)

* feat: change ExpireableMap definition (#1563)

* [Fix] Translation issue & Paintbrush GUI (#1564)

* fix: translation issue

* i18n: update reverser key binds

* [Feature] LightFlare changes (#1559)

* feat: add direction check ability for headlight

* feat: add config for light texture render

* feat: add ability to invert headlight direction

* update base.caml

* snowLayers property for rolling stock (#1567)

* feat: Add new property "snow_layers" to rolling stock

* feat: Add desired snow layer height to cleanSnow() method

* fix translation error (#1570)

* i18n: update zh-cn file (#1573)

* Allow golden spike to modify slope height (#1566)

higher or lower than the starting block

* Add cargo load command (#1569)

Usage: /immersiverailroading cargoLoad <radius>

Will use whatever item is held to fill the empty slots in the car

* [feat] Added semi-transparent parts (#1556)

* [feat] Added partial-transparency

* [feat] Added animations to partial-transparent parts

* [cleanup] removed unnecessary imports

* [fix] Replaced filter streams for better performance

* [feat] Added Config option to disable the rendering of semi-transparent parts

* [fix] Minor improvements of the performance of ModelState.render

* Cleanup ModelState to make changes like transparency easier

---------


* feat: add excludeStandaloneWagons to optimize performance (#1576)

* revert #1555 (#1582)

* fix structure (#1580)

* fix: unexpected fallthrough (#1586)

* [Fix & Feature] Golden spike changes (#1572)

* feat: refactor golden spike behavior on horizontal

* fix issue#1455

* Customizable decimals on GUI readouts (#1512)

* Removed decimals from GUI speed readout

* Added reverser controls to en_us.lang

* remove translation key as added in #1554

* feat: add decimal settings to gui files

* fix: correct misspell

* change `SPEED` default decimals

* fix

* minor optimize

* Revert default GUIs to 2 decimals precision

* Add missing space before comment

---------


* [Feature] Advanced track model definition (#1565)

* feat: new track model definition

* fix: add back switch compatibility for rail model

* refactor: rename fields

* feat: add TABLE component for turntable and transfer table

* feat: rewrite order implement

* refactor

* refactor

* feat: rewrite random implement

* feat: optimize random calculation

* fix: a null pointer and docs

* solve conflicts

* feat: change ExpireableMap definition (#7)

* feat: rewrite definition format and clean-up

* fix: a potential NPE

* continuous fix

* fix: optimize

* fix: use isTable()

* fix a potential NPE

* solve conflict

* minor optimize

* [Feature] Add ability to customize diesel fuel by definition (#1577)

* feat: add ability to customize diesel fuel by definition

* docs: update diesel.caml

* feat: add `whitelistBlocks` to prevent destruction to certain blocks (#1579)

* feat: add `whitelistBlocks` to prevent destruction to certain blocks

* fix: change implement

* refactor: change config entry name

* [Feature] SI Units (#1585)

* feat: add some SI unit

* feat: replace tooltip

* fix

* fixes about SI

* fix: unexpected fallthrough

* refactor

* refactor: introduce const

* refactor: extract const and remove Math.ceil

* fix: NPE (#1587)

* Remove decimals from tooltips (#1590)

Remove 2 decimal figures added to locomotive tooltips in 376a848

* Fix SI units and add power_hp as alternative to horsepower (#1589)

* Fix SI units and add power_hp as alternative to horsepower

* Fixed rounding of power_w

* fix: decimal digits and add docs

* Change psiToKPa to exactly match inverse of psiToKPa to prevent potential calculation drift

* Added bar as definable max pressure unit and implemented new style for max_pressure definitions

kept max_psi, but changed max_kpa to max_pressure_kpa, since I don't think anyone has used it yet, additionally, max_pressure_psi and max_pressure_bar also exist.

* Reverted back to divsion by 1000 instead of multiplication by 0.001, as it reduces floating point imprecision

* Changed default caml files to make them more clear and include new definitions

---------


* fix: track with too many segments won't work (#1591)

* Added bar, adjusted hp, added ps, adjusted defaults, organized unit library (#1592)

Added new units: bar and ps, adjusted horsepower and power_hp to be imperial horsepower (was metric in the past few commits), fixed some capitalization mistakes, and reordered the orders of the config options for selecting units in the config as well as adjusting the defaults to all be metric in accordance with discussions on discord.

* docs: update changelog

* [Refactor] VecYawPitch cleanup (#1595)

* docs: update changelog

* refactor: move VecYawPitch out of BuilderBase

* refactor: remap PosStep to VecYawPitch

* feat: rename VecYawPitch and add roll axis

* fix: incorrect call of getLength()

* fix: slope without golden spike won't work (#1597)

* [I18n] Add missing entries for lang files & update ja_jp.lang(Translated by @FS_FADE) (#1596)

* i18n: add missing entries for lang files

* i18n: update zh_cn.lang

* i18n: update ja_jp.lang(Translated by @FS_FADE)

---------

+++ src/main/java/cam72cam/immersiverailroading/entity/LocomotiveSteam.java
@@ -286,7 +286,7 @@ public class LocomotiveSteam extends Locomotive {
 			}
 			
 			// Pressure relief valve
-			int maxPSI = this.getDefinition().getMaxPSI(gauge);
+			int maxPSI = (int) this.getDefinition().getMaxPSI(gauge);
 			pressureValve = boilerPressure > maxPSI;
 			if (boilerPressure > maxPSI) {
 				waterUsed += boilerPressure - maxPSI;
